Move preload graceful method tests to a more proper location in config manager tests, and fix.
